+++ coreTests/UnknownGrammarProgram.test.ts
@@ -47,7 +47,7 @@ langs.forEach((name: string) => {
     // Act
     const inferredPrefixGrammarCode = new UnknownGrammarProgram(sampleCode).inferGrammarFileForAKeywordLanguage("foobar")
     const inferredPrefixGrammarProgram = new jtree.GrammarProgram(inferredPrefixGrammarCode)
-    const rootProgramConstructor = inferredPrefixGrammarProgram.getRootConstructor()
+    const rootProgramConstructor = inferredPrefixGrammarProgram.compileAndReturnRootConstructor()
     const programParsedWithInferredGrammar = new rootProgramConstructor(sampleCode)
 
     // Assert